+++ jest.config.ts
@@ -13,6 +13,7 @@ const config: Config.InitialOptions = {
     "@nimbus-ds/menu": "<rootDir>/packages/react/Menu/src",
     "@nimbus-ds/page": "<rootDir>/packages/react/Page/src",
     "@nimbus-ds/layout": "<rootDir>/packages/react/Layout/src",
+    "@nimbus-ds/data-list": "<rootDir>/packages/react/DataList/src",
   },
 };