+++ packages/react/InteractiveList/src/components/InteractiveListCheckboxItem/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListCheckboxItem } from "./InteractiveListCheckboxItem";
+
+export { InteractiveListCheckboxItem } from "./InteractiveListCheckboxItem";
+export default InteractiveListCheckboxItem;