+++ packages/react/Menu/src/components/MenuHeader/menuHeader.types.ts
@@ -1,6 +1,7 @@
 import { HTMLAttributes, ReactNode } from "react";
 
-export interface MenuHeaderProps extends HTMLAttributes<HTMLElement> {
+export interface MenuHeaderProps
+  extends Omit<HTMLAttributes<HTMLElement>, "color"> {
   /** Content of the menu header */
   children: ReactNode;
 }