+++ packages/react/Page/src/components/PageBody/pageBody.types.ts
@@ -1,6 +1,7 @@
 import { ReactNode, HTMLAttributes } from "react";
 
-export interface PageBodyProps extends HTMLAttributes<HTMLElement> {
+export interface PageBodyProps
+  extends Omit<HTMLAttributes<HTMLElement>, "color"> {
   /** Content of the page body */
   children: ReactNode;
 }